S15.d.4: extract up.ts sandbox bring-up (1182 → 766 LOC) — caps S15.d ✅
Final sub-slice of the up.ts multi-PR sub-train. Extract the sandbox
bring-up phase (Step 7: federated credentials + Foundry RBAC +
ClawSandbox CR; Step 8: wait for Running + WebUI port-forward;
deployment summary + saveContext()) to cli/src/commands/up/sandbox_bringup.ts.

Caller in up.ts shrinks to a 9-line dispatch:
  const { bringUpSandbox } = await import("./up/sandbox_bringup.js");
  await bringUpSandbox({
    options,
    baseName, rg,
    acrLoginServer, foundryEndpoint, openAiEndpoint, kvName,
    wiClientId, imdsClientId,
    repoRoot, stepper,
    registryMode, globalRegistryUrl, globalRelayUrl,
  });

Cumulative S15.d delta:
  pre-S15.d: 1849
  d.1: 1660 (-189) — fast_upgrade.ts
  d.2: 1296 (-364) — preflight.ts
  d.3: 1182 (-114) — agentmesh_deploy.ts
  d.4:  766 (-416) — sandbox_bringup.ts
  §4.2 cap of 800 achieved ✅

Verification:
- up.ts: 1182 → 766 LOC
- new up/sandbox_bringup.ts: 482 LOC
- tsc --noEmit / lint (27 warnings, baseline) / build: clean
- npm test: 454 passed; 2 skipped pre-existing

Removed unused imports from up.ts: { section, kvLine, checkLine }
from stepper.js and { saveContext } from config.js (now used only
inside the new module).

No behavioral change: sandbox bring-up body moved verbatim. Federated
credential audience (api://AzureADTokenExchange), Foundry RBAC role
IDs (Azure AI User 53ca6127-…, Cog Svc OpenAI User 5e0bd9bd-…),
ClawSandbox CR shape (trustThreshold: 500, contentSafety + promptShields:
true, defaultDeny + approvalRequired + learnEgress: true), and saveContext()
payload all preserved verbatim.

Audit: docs/security-audits/2026-04-29-phase2-hotspot-up-cli-d4.md
